pyproject.toml
[tool.poetry]
name = "wallet-transaction-api"
version = "0.1.0"
description = ""
authors = ["Artem Getashvili <artemget98@yandex.ru>"]
readme = "README.md"
packages = [{include = "wallet_transaction_api"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework-jsonapi = "^7.0.2"
mysqlclient = "^2.2.4"
django-environ = "^0.11.2"
alembic = "^1.13.2"
django-filter = "^24.3"
pytest-django = "^4.8.0"
pytest-cov = "^5.0.0"
loguru = "^0.7.2"
drf-yasg = "^1.21.7"
django-redis = "^5.4.0"
pre-commit = "^3.8.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
isort = "^5.13.2"
pre-commit = "^3.8.0"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
